requirements.txt
hydra-core >= 1.3.2
opencv-python >= 4.5.5.62
scipy >= 1.3.0
scikit-image >= 0.13.0
av >= 10.0.0
six >= 1.16.0
mediapipe==0.10.3
python-dotenv==1.0.0
jellyfish==1.0.0
openai==0.28.0
moviepy==1.0.3
num2words==0.5.12
scikit-learn==1.3.0
python-Levenshtein==0.21.1
google-cloud-speech==2.21.0
SpeechRecognition==3.10.0
Flask==2.3.3